Change permission of httpd.pid file.
Add read, write and execute rights for group. As the owner of the pid file is root, users get a problem if they have set their umask to 067, cause they no longer have the right to read the file and thus the included process id. As such the server will not be killed.